gitbak

+++ bench/label/add_level.cpp
@@ -7,7 +7,7 @@
 
 B63_BASELINE(add_alllevel, nn) {
 	kAryLabel v1{}, v2{}, v3{};
-	uint64_t k_lower=0, k_higher=0;
+	uint32_t k_lower=0, k_higher=0;
 	B63_SUSPEND {
 		for (uint64_t i = 0; i < k; ++i) {
 			v1.data()[i] = i;
@@ -33,7 +33,7 @@ B63_BASELINE(add_alllevel, nn) {
 
 B63_BENCHMARK(add_level1, nn) {
 	kAryLabel v1{}, v2{}, v3{};
-	uint64_t k_lower=0, k_higher=0;
+	uint32_t k_lower=0, k_higher=0;
 	B63_SUSPEND {
 		for (uint64_t i = 0; i < k; ++i) {
 			v1.data()[i] = i;
@@ -59,7 +59,7 @@ B63_BENCHMARK(add_level1, nn) {
 
 B63_BENCHMARK(add_level2, nn) {
 	kAryLabel v1{}, v2{}, v3{};
-	uint64_t k_lower=0, k_higher=0;
+	uint32_t k_lower=0, k_higher=0;
 	B63_SUSPEND {
 		for (uint64_t i = 0; i < k; ++i) {
 			v1.data()[i] = i;
@@ -85,7 +85,7 @@ B63_BENCHMARK(add_level2, nn) {
 
 B63_BENCHMARK(add_level3, nn) {
 	kAryLabel v1{}, v2{}, v3{};
-	uint64_t k_lower, k_higher;
+	uint32_t k_lower, k_higher;
 	B63_SUSPEND {
 		for (uint64_t i = 0; i < k; ++i) {
 			v1.data()[i] = i;
